+++ hosts/hermes/darwin-configuration.nix
@@ -3,6 +3,10 @@
 {
   networking.knownNetworkServices = [ "Wi-Fi" ];
 
+  system.activationScripts.extraActivation.text = ''
+    nvram StartupMute=%01
+  '';
+
   nix.registry.ln.to = {
     type = "git";
     url = "file:///Users/${user}/Code/nixpkgs";