+++ modules/vm.nix
@@ -27,7 +27,7 @@
     };
 
   # WORKAROUND: { osConfig ? { }, ... }: fails when using `home-manager build`
-  hmModule = { inputs, lib, pkgs, ... }@args:
+  hmModule = { inputs, pkgs, lib, ... }@args:
     let inherit (lib) hasAttrByPath mkIf mkVMOverride;
     in mkIf (hasAttrByPath [ "osConfig" "virtualisation" "diskImage" ] args) {
       services.polybar.config."bar/centre".monitor = mkVMOverride "Virtual-1";